Split pages: - add additional pages (buy, checkout) - delete BuyForm - simplify order process - code refactoring
