+++ .eslintrc.json
@@ -1,6 +1,6 @@
 {
   "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"],
+  "plugins": ["@typescript-eslint", "prettier"],
   "extends": [
     "eslint:recommended",
     "plugin:react/recommended",
@@ -45,6 +45,7 @@
         "shorthandLast": true,
         "ignoreCase": false
       }
-    ]
+    ],
+    "prettier/prettier": 2
   }
 }
