--- components/BuyForm/BuyForm.module.scss
@@ -1,61 +1,0 @@
-.textField {
-  background-color: #fff;
-}
-
-.radio {
-  padding: 5px;
-}
-
-.controlLabelRoot {
-  margin-left: -7px;
-}
-
-.controlLabel {
-  @apply text-xl;
-}
-
-.radioGroupLabel {
-  @apply text-xl text-gray-700;
-  margin-bottom: 10px;
-}
-
-.countInput {
-  width: 30px;
-  text-align: center;
-
-  &::-webkit-inner-spin-button,
-  &::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-}
-
-.countInputErrorText {
-  color: #f44336;
-}
-
-.resultTable {
-  & th {
-    padding-right: 20px;
-    text-align: left;
-  }
-
-  & tr:last-child {
-    & th,
-    td {
-      padding-top: 20px;
-    }
-  }
-}
-
-.submit {
-  @apply text-center;
-}
-
-.submitButton {
-  @apply bg-green-600 text-white font-bold py-2 px-4 rounded;
-
-  &:hover {
-    @apply bg-green-700;
-  }
-}
