--- components/delivery/forms/CDEKDeliveryForm/CDEKDeliveryForm.module.scss
@@ -1,20 +1,0 @@
-.radio {
-  padding: 5px;
-}
-
-.controlLabelRoot {
-  margin-left: -7px;
-}
-
-.controlLabel {
-  @apply text-xl;
-}
-
-.radioGroupLabel {
-  @apply text-xl text-gray-700;
-  margin-bottom: 10px;
-}
-
-.textField {
-  @apply bg-white;
-}
