+++ components/delivery/forms/KazPostDeliveryForm.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
-import { TextField } from '@material-ui/core';
 
-import styles from './KazPostDeliveryForm.module.scss';
+import { FormLabel, TextField } from '@material-ui/core';
 
 const kazPostDeliveryForm = (): JSX.Element => {
   const { register, errors } = useFormContext();
@@ -11,11 +10,10 @@ const kazPostDeliveryForm = (): JSX.Element => {
     <>
       <div className="flex flex-wrap mb-6">
         <div className="w-full md:w-2/3 md:pr-3 mb-6 md:mb-0">
-          <label className="block text-gray-700" htmlFor="grid-delivery-city">
+          <FormLabel focused={false} required>
             Населённый пункт
-          </label>
+          </FormLabel>
           <TextField
-            className={styles.textField}
             error={!!errors?.city}
             helperText={errors?.city?.message}
             inputRef={register({
@@ -29,11 +27,10 @@ const kazPostDeliveryForm = (): JSX.Element => {
           />
         </div>
         <div className="w-full md:w-1/3 ">
-          <label className="block text-gray-700" htmlFor="grid-delivery-zip">
+          <FormLabel focused={false} required>
             Почтовый индекс
-          </label>
+          </FormLabel>
           <TextField
-            className={styles.textField}
             error={!!errors?.zip}
             helperText={errors?.zip?.message}
             inputRef={register({
@@ -47,12 +44,11 @@ const kazPostDeliveryForm = (): JSX.Element => {
           />
         </div>
       </div>
-      <div className="w-full mb-6">
-        <label className="block text-gray-700" htmlFor="grid-delivery-address">
+      <div className="w-full">
+        <FormLabel focused={false} required>
           Адрес получателя
-        </label>
+        </FormLabel>
         <TextField
-          className={styles.textField}
           error={!!errors?.address}
           helperText={errors?.address?.message}
           inputRef={register({
