--- components/delivery/forms/KazPostDeliveryForm/KazPostDeliveryForm.module.scss
@@ -1,3 +1,0 @@
-.textField {
-  @apply bg-white;
-}
