+++ components/layouts/page.tsx
@@ -4,16 +4,20 @@ import Header from 'components/Header';
 import Footer from 'components/Footer';
 
 interface Props {
-  children?: JSX.Element[];
+  children?: string | JSX.Element[] | JSX.Element;
 }
 
 const PageLayout = ({ children }: Props): JSX.Element => {
   return (
-    <div className="flex flex-col min-h-screen text-gray-700">
+    <div className="flex flex-col min-h-screen text-gray-700 h-full">
       <Header />
-      <section className="flex flex-col flex-grow bg-white border-b py-8">
-        {children}
-      </section>
+
+      <div className="flex-grow bg-white py-8">
+        <div className="container max-w-screen-xl mx-auto h-full">
+          {children}
+        </div>
+      </div>
+
       <Footer />
     </div>
   );
