+++ interfaces/OrderForm.interface.ts
@@ -1,7 +1,7 @@
 import { ICustomer } from './Customer.interface';
 import { ICDEKCityItem } from './CdekCityItem.interface';
 
-export interface IBuyForm extends ICustomer {
+export interface IOrderForm extends ICustomer {
   city: ICDEKCityItem | string;
   deliveryCompany: string;
   zip: string;
