+++ styles/main.css
@@ -2,48 +2,19 @@
 @tailwind components;
 @tailwind utilities;
 
+html {
+  height: 100%;
+}
 body {
   font-family: 'Roboto', sans-serif;
-}
-.gradient {
-  background: linear-gradient(90deg, #2f6844 0%, #71bf8f 100%);
-}
-
-.buyButton {
-  box-shadow: 0 1px 0 0 #fff6af;
-  background: #ffec64 linear-gradient(to bottom, #ffec64 5%, #ffab23 100%);
-  border-radius: 6px;
-  border: 1px solid #ffaa22;
-  display: inline-block;
-  cursor: pointer;
-  color: #333333;
-  font-size: 15px;
-  font-weight: bold;
-  padding: 6px 24px;
-  text-decoration: none;
-  text-shadow: 0px 1px 0px #ffee66;
-}
-.buyButton:hover {
-  background: #ffab23 linear-gradient(to bottom, #ffab23 5%, #ffec64 100%);
-}
-.buyButton:active {
-  position: relative;
-  top: 1px;
-}
-
-.note {
-  background: #d4f9e214;
-  border: 1px solid #306a454f;
-  margin: 1rem 0;
-  padding: 1rem;
   height: 100%;
 }
 
-@media (min-width: 640px) {
-  .note {
-    margin: 1rem;
-    padding: 2rem 3rem 3rem;
-  }
+#__next {
+  height: 100%;
+}
+.gradient {
+  background: linear-gradient(90deg, #2f6844 0%, #71bf8f 100%);
 }
 
 button:focus {
